feat(52.2-02): persisted-mode payload guard, media reference, runtime byte assertion
- efxPaintPackage.ts: FrameMediaReference + fail-closed parseFrameMediaReference
  (frames/ prefix, 64 lower-case hex digest, positive width/height)
- physicsPaintRotoPhysicalModel.ts: optional media carrier on the payload,
  PhysicPaintRotoPayloadMode ('runtime' | 'reference-only'), the persisted
  allowlist without the raster key, buildPhysicPaintRotoPayloadContentToken
  (byte token for runtime, media digest for reference-only) and
  requirePhysicPaintRotoInlineBytes for every runtime pixel projection
- efxPaintDocumentParsers.ts: payloadMode threaded through parseEfxPaintDocument
  and parseInternalPaintTrack, defaulting to 'runtime'
- efxPaintPersistence.ts: loadEfxPaintDocuments is the single non-test call site
  that selects 'reference-only' (the on-disk READ door)
- store, bridge, resolver, both Roto hooks, onion preview and Studio now compare
  payload content by token and assert the inline carrier where pixels are read
- parser suite: named traversal/digest/allowlist refusals, the bytes-only runtime
  default, four group-override cases and the on-disk-door mode selection

+++ app/src/lib/efxPaintPersistence.test.ts
@@ -213,7 +213,9 @@ describe('saveEfxPaintDocumentsWithProjectWrite / loadEfxPaintDocuments', () => 
     // Load decodes base64 back to a Uint8Array.
     const loaded = await loadEfxPaintDocuments('/project', payload);
     const restored = loaded.get('layer-bytes')!.document;
-    const restoredBytes = restored.tracks[0].rotoPhysical!.realKeyRecords[0].payload.bytes;
+    // 52.2-02: the payload's raster carrier is optional on the shared record
+    // shape, so the inline-bytes assertion is the one that pins it.
+    const restoredBytes = restored.tracks[0].rotoPhysical!.realKeyRecords[0].payload.bytes as Uint8Array;
     expect(restoredBytes).toBeInstanceOf(Uint8Array);
     expect(Array.from(restoredBytes)).toEqual(Array.from(bytes));
   });
